fix state deadlock

+++ MusuemHeistSemaphores/run.sh
@@ -1,3 +1,11 @@
+#!/bin/bash
+
+if [ ! -n "$1" ];
+then
+	printf "usage:\n sh run.sh <log_file>\n"
+	exit 1	
+fi
+
 find . -name "*.class" -type f -print0 | xargs -0 /bin/rm -f
-javac MusuemHeist.java
-java -Djava.util.logging.SimpleFormatter.format='%1$tY-%1$tm-%1$td %1$tH:%1$tM:%1$tS %4$s %2$s %5$s%6$s%n' MusuemHeist
+javac MuseumHeist.java
+java  MuseumHeist $1
